pyproject.toml
[project]
name = "econatlas"
version = "0.1.0"
description = "CLI crawler for economics journals with DeepSeek translations"
readme = "README.md"
authors = [
    { name = "Icarus603", email = "177302395+Icarus603@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0.11",
    "httpx>=0.27.0",
    "langdetect>=1.0.9",
    "pydantic>=2.7.0",
    "python-dateutil>=2.9.0",
    "python-dotenv>=1.0.1",
    "python-slugify>=8.0.4",
    "playwright>=1.45.0",
    "typer>=0.12.3",
    "beautifulsoup4>=4.14.2",
]

[project.scripts]
econ-atlas = "econatlas.cli:app"

[build-system]
requires = ["uv_build>=0.9.9,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.10.0",
    "pytest>=8.2.0",
    "ruff>=0.5.0",
    "types-python-dateutil>=2.9.0.20240316",
]
